requirements.txt
# Core Django and API framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0
django-ratelimit==4.1.0
django-extensions==3.2.3

# Database and ORM
psycopg2-binary==2.9.7
django-environ==0.11.2

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Task queue and caching
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Authentication and security
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.1.2
tenacity==8.2.3

# External service clients
openai==2.2.0
pinecone-client==2.2.4
stripe==7.8.0
boto3==1.34.0

# File processing
pypdf2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
beautifulsoup4==4.12.2
requests==2.31.0
youtube-transcript-api==0.6.1

# Security and virus scanning
# pyclamd==0.4.0  # Uncomment when ClamAV integration is needed

# Text processing and embeddings
langchain==0.0.340
tiktoken==0.5.2
sentence-transformers>=2.2.0  # For semantic reranking (optional - gracefully degrades)

# Monitoring and observability
structlog==23.2.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-django==0.43b0
opentelemetry-instrumentation-celery==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-requests==0.43b0
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
responses==0.24.1

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
safety==2.3.4

# Development tools
ipython==8.18.1
django-debug-toolbar==4.2.0

# Production server
gunicorn==21.2.0
uvicorn==0.24.0

# Documentation
drf-spectacular==0.26.5